Test kframework 8

+++ rv-predict-engine/src/main/java/rvpredict/trace/MemoryAccessEvent.java
@@ -33,7 +33,7 @@ package rvpredict.trace;
  */
 public abstract class MemoryAccessEvent extends InitOrAccessEvent {
 
-    protected MemoryAccessEvent(long GID, long TID, int ID, EventType type, long objectHashCode,
+    protected MemoryAccessEvent(long GID, long TID, int ID, EventType type, int objectHashCode,
             int index, long value) {
         super(GID, TID, ID, type, objectHashCode, index, value);
         assert type == EventType.READ || type == EventType.WRITE;
